refactor(ui): TE-1270 renamed .interface suffix to .interfaces for standardisation
