+++ thirdeye-ui/src/app/components/subscription-group-view/alert-associations-view-table/alert-associations-view-table.interfaces.ts
