+++ thirdeye-ui/src/app/components/subscription-group-wizard/alerts-dimensions/alert-dimension-table/alert-dimension-table.interfaces.ts
@@ -14,15 +14,12 @@
  */
 
 import { Alert } from "../../../../rest/dto/alert.interfaces";
-import { Association } from "../../subscription-group-wizard.interface";
+import { EnumerationItem } from "../../../../rest/dto/enumeration-item.interfaces";
+import { Association } from "../../subscription-group-wizard.interfaces";
 
-export interface DimensionRow {
-    id: string;
-    name: string;
-}
-
-export interface AddDimensionDialogProps {
-    alerts: Alert[];
+export interface AlertDimensionTableProps {
+    selectedAlert: Alert;
     associations: Association[];
-    updateAssociations: (p: Association[]) => void;
+    onChangeAssociations: (p: string[]) => void;
+    enumerationItemsForAlert: EnumerationItem[];
 }
