+++ thirdeye-ui/src/app/components/subscription-group-wizard/subscription-group-details/properties-form/properties-form.interfaces.ts
