+++ thirdeye-ui/src/app/components/subscription-group-wizard/subscription-group-details/recipient-details/groups-renderer/sendgrid-email-review/sendgrid-email-review.interfaces.ts
