+++ thirdeye-ui/src/app/components/subscription-group-wizard/subscription-group-details/subscription-group-details.interfaces.ts
